ApiUseCaseにモジュールを追加
