+++ app/src/main/kotlin/jp/co/yumemi/android/codeCheck/di/ui/OneViewModel.kt
@@ -10,6 +10,7 @@ import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
 import jp.co.yumemi.android.codeCheck.di.data.GitItem
 import jp.co.yumemi.android.codeCheck.di.domain.api.GetGitApiUseCase
+import jp.co.yumemi.android.codeCheck.di.domain.api.GetGitApiUseCaseImpl
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.runBlocking
 import javax.inject.Inject
@@ -18,15 +19,15 @@ import javax.inject.Inject
  * TwoFragment で使う
  */
 @HiltViewModel
-class OneViewModel @Inject constructor() : ViewModel() {
+class OneViewModel @Inject constructor(
+    private val getGitApiUseCase: GetGitApiUseCase
+) : ViewModel() {
 
     //
     private var languageFormat: String = ""
 
     private var _searchResult: MutableLiveData<List<GitItem>> = MutableLiveData()
     val searchResult: LiveData<List<GitItem>> get() = _searchResult
-
-    private val useGitApi = GetGitApiUseCase()
 
     fun setLanguageFormat(text: String) {
         languageFormat = text
@@ -35,7 +36,7 @@ class OneViewModel @Inject constructor() : ViewModel() {
     // 入力されたTextでRepositoryを検索
     fun searchResults(inputText: String) = runBlocking {
         viewModelScope.launch {
-            _searchResult.value = useGitApi.getSearchApi(inputText, languageFormat)
+            _searchResult.value = getGitApiUseCase.getSearchApi(inputText, languageFormat)
         }
     }
 }
